services/producer/requirements.txt
kafka-python==2.0.2
pydantic==2.7.4